// src/runtime/types.ts
export type TransformLink = (
        href: string,
        children: Array<ElementContent>,
        title: string | null
) => string

export type TransformImage = (
        src: string,
        alt: string,
        title: string | null
) => string

export type TransformLinkTarget = (
        href: string,
        children: Array<ElementContent>,
        title: string | null
) => TransformLinkTargetType | undefined

export type TransformLinkTargetType =
        | '_self'
        | '_blank'
        | '_parent'
        | '_top'
        | (string & {})

export type TransformOptions = {
        sourcePos?: boolean
        rawSourcePos?: boolean
        skipHtml?: boolean
        includeElementIndex?: boolean
        transformLinkUri?: false | TransformLink | null
        transformImageUri?: TransformImage
        linkTarget?: TransformLinkTarget | TransformLinkTargetType
        components?: Components
}

export type Context = {
        options: TransformOptions
        schema: Schema
        listDepth: number
}
// ...
export type Components = Partial<
        Omit<NormalComponents, keyof SpecialComponents>
> &
        Partial<SpecialComponents>